implement server-driven execution of alternate engine chunks
