+++ src/cpp/session/modules/rmarkdown/SessionRMarkdown.cpp
@@ -15,7 +15,6 @@
 
 #include "SessionRMarkdown.hpp"
 #include "SessionRmdNotebook.hpp"
-#include "NotebookCache.hpp"
 #include "../SessionHTMLPreview.hpp"
 #include "../build/SessionBuildErrors.hpp"
 
@@ -46,7 +45,6 @@
 #include <session/projects/SessionProjects.hpp>
 
 #include "RMarkdownPresentation.hpp"
-#include "SessionExecuteChunkOperation.hpp"
 
 #define kRmdOutput "rmd_output"
 #define kRmdOutputLocation "/" kRmdOutput "/"
@@ -1324,369 +1322,6 @@ Error maybeCopyWebsiteAsset(const json::JsonRpcRequest& request,
    return Success();
 }
 
-class ChunkExecCompletedScope
-{
-public:
-   ChunkExecCompletedScope(const std::string& docId,
-                           const std::string& chunkId)
-      : docId_(docId), chunkId_(chunkId)
-   {
-   }
-   
-   ~ChunkExecCompletedScope()
-   {
-      notebook::events().onChunkExecCompleted(
-               docId_,
-               chunkId_,
-               notebook::notebookCtxId());
-   }
-   
-private:
-   std::string docId_;
-   std::string chunkId_;
-};
-
-void chunkConsoleOutputHandler(module_context::ConsoleOutputType type,
-                               const std::string& output,
-                               const FilePath& targetPath)
-{
-   using namespace module_context;
-   
-   Error error = notebook::appendConsoleOutput(
-            type == ConsoleOutputNormal ? kChunkConsoleOutput : kChunkConsoleError,
-            output,
-            targetPath);
-   if (error)
-      LOG_ERROR(error);
-}
-
-Error prepareCacheConsoleOutputFile(const std::string& docId,
-                                    const std::string& chunkId,
-                                    const std::string& nbCtxId,
-                                    FilePath* pChunkOutputFile)
-{
-   // forward declare error
-   Error error;
-   
-   // prepare chunk directory
-   FilePath cachePath = notebook::chunkOutputPath(
-            docId,
-            chunkId,
-            notebook::ContextExact);
-   
-   error = cachePath.removeIfExists();
-   if (error)
-      return error;
-   
-   error = cachePath.ensureDirectory();
-   if (error)
-      return error;
-   
-   // prepare cache console output file
-   *pChunkOutputFile =
-         notebook::chunkOutputFile(docId, chunkId, nbCtxId, kChunkOutputText);
-   
-   return Success();
-}
-
-Error executeRcppEngineChunk(const std::string& docId,
-                             const std::string& chunkId,
-                             const std::string& nbCtxId,
-                             const std::string& code,
-                             const std::map<std::string, std::string>& options,
-                             json::JsonRpcResponse* pResponse)
-{
-   // forward declare error
-   Error error;
-   
-   // always ensure we emit a 'execution complete' event on exit
-   ChunkExecCompletedScope execScope(docId, chunkId);
-   
-   // prepare cache output file (use tempfile on failure)
-   FilePath targetPath = module_context::tempFile("rcpp-cache", "");
-   error = prepareCacheConsoleOutputFile(docId, chunkId, nbCtxId, &targetPath);
-   if (error)
-      LOG_ERROR(error);
-   
-   // capture console output, error
-   boost::signals::scoped_connection consoleHandler =
-         module_context::events().onConsoleOutput.connect(
-            boost::bind(chunkConsoleOutputHandler,
-                        _1,
-                        _2,
-                        targetPath));
-
-   // call Rcpp::sourceCpp on code
-   std::string escaped = boost::regex_replace(
-            code,
-            boost::regex("(\\\\|\")"),
-            "\\\\$1");
-   
-   std::string execCode =
-         "Rcpp::sourceCpp(code = \"" + escaped + "\")";
-   
-   // write input code to cache
-   error = notebook::appendConsoleOutput(
-            kChunkConsoleInput,
-            code,
-            targetPath);
-   if (error)
-      LOG_ERROR(error);
-   
-   // execute code (output captured on success; on failure we
-   // explicitly forward the error message returned)
-   error = r::exec::executeString(execCode);
-   if (error)
-   {
-      chunkConsoleOutputHandler(module_context::ConsoleOutputError,
-                                r::endUserErrorMessage(error),
-                                targetPath);
-   }
-
-   // forward success / failure to chunk
-   notebook::enqueueChunkOutput(
-            docId,
-            chunkId,
-            nbCtxId,
-            kChunkOutputText,
-            targetPath);
-   
-   return error;
-}
-
-void reportChunkExecutionError(const std::string& docId,
-                               const std::string& chunkId,
-                               const std::string& nbCtxId,
-                               const std::string& message,
-                               const FilePath& targetPath)
-{
-   // emit chunk error
-   chunkConsoleOutputHandler(
-            module_context::ConsoleOutputError,
-            message,
-            targetPath);
-   
-   // forward failure to chunk
-   notebook::enqueueChunkOutput(
-            docId,
-            chunkId,
-            nbCtxId,
-            kChunkOutputText,
-            targetPath);
-}
-
-void reportStanExecutionError(const std::string& docId,
-                              const std::string& chunkId,
-                              const std::string& nbCtxId,
-                              const FilePath& targetPath)
-{
-   std::string message =
-         "engine.opts$x must be a character string providing a "
-         "name for the returned `stanmodel` object";
-   reportChunkExecutionError(docId, chunkId, nbCtxId, message, targetPath);
-}
-
-
-Error executeStanEngineChunk(const std::string& docId,
-                             const std::string& chunkId,
-                             const std::string& nbCtxId,
-                             const std::string& code,
-                             const std::map<std::string, std::string>& options,
-                             json::JsonRpcResponse* pResponse)
-{
-   // forward-declare error
-   Error error;
-   
-   // ensure we always emit an execution complete event on exit
-   ChunkExecCompletedScope execScope(docId, chunkId);
-   
-   // prepare console output file -- use tempfile on failure
-   FilePath targetPath = module_context::tempFile("stan-cache-", "");
-   error = prepareCacheConsoleOutputFile(docId, chunkId, nbCtxId, &targetPath);
-   if (error)
-      LOG_ERROR(error);
-   
-   // capture console output, error
-   boost::signals::scoped_connection consoleHandler =
-         module_context::events().onConsoleOutput.connect(
-            boost::bind(chunkConsoleOutputHandler,
-                        _1,
-                        _2,
-                        targetPath)); 
-   
-   // write code to file
-   FilePath tempFile = module_context::tempFile("stan-", ".stan");
-   error = writeStringToFile(tempFile, code);
-   if (error)
-   {
-      reportChunkExecutionError(
-               docId,
-               chunkId,
-               nbCtxId,
-               r::endUserErrorMessage(error),
-               targetPath);
-      LOG_ERROR(error);
-      return Success();
-   }
-   RemoveOnExitScope removeOnExitScope(tempFile, ERROR_LOCATION);
-   
-   // ensure existence of 'engine.opts' with 'x' parameter
-   if (!options.count("engine.opts"))
-   {
-      reportStanExecutionError(docId, chunkId, nbCtxId, targetPath);
-      return Success();
-   }
-   
-   // evaluate engine options (so we can pass them through to stan call)
-   r::sexp::Protect protect;
-   SEXP engineOptsSEXP = R_NilValue;
-   error = r::exec::evaluateString(
-            options.at("engine.opts"),
-            &engineOptsSEXP,
-            &protect);
-   
-   if (error)
-   {
-      reportStanExecutionError(docId, chunkId, nbCtxId, targetPath);
-      return Success();
-   }
-   
-   // construct call to 'stan_model'
-   r::exec::RFunction fStanEngine("rstan:::stan_model");
-   std::vector<std::string> engineOptsNames;
-   error = r::sexp::getNames(engineOptsSEXP, &engineOptsNames);
-   if (error)
-   {
-      reportStanExecutionError(docId, chunkId, nbCtxId, targetPath);
-      return Success();
-   }
-   
-   // build parameters
-   std::string modelName;
-   for (std::size_t i = 0, n = r::sexp::length(engineOptsSEXP); i < n; ++i)
-   {
-      // skip 'x' engine option (this is the variable we wish to assign to
-      // after evaluating the stan model)
-      if (engineOptsNames[i] == "x")
-      {
-         modelName = r::sexp::asString(VECTOR_ELT(engineOptsSEXP, i));
-         continue;
-      }
-      
-      fStanEngine.addParam(
-               engineOptsNames[i],
-               VECTOR_ELT(engineOptsSEXP, i));
-   }
-   
-   // if no model name was set, return error message
-   if (modelName.empty())
-   {
-      reportStanExecutionError(docId, chunkId, nbCtxId, targetPath);
-      return Success();
-   }
-   
-   // if the 'file' option was not set, set it explicitly
-   if (!core::algorithm::contains(engineOptsNames, "file"))
-      fStanEngine.addParam("file", tempFile.absolutePath());
-   
-   // evaluate stan_model call
-   SEXP stanModelSEXP = R_NilValue;
-   error = fStanEngine.call(&stanModelSEXP, &protect);
-   if (error)
-   {
-      std::string msg = r::endUserErrorMessage(error);
-      reportChunkExecutionError(docId, chunkId, nbCtxId, msg, targetPath);
-      return Success();
-   }
-   
-   // assign in global env on success
-   if (stanModelSEXP != R_NilValue)
-   {
-      r::exec::RFunction assign("base:::assign");
-      assign.addParam("x", modelName);
-      assign.addParam("value", stanModelSEXP);
-      error = assign.call();
-      if (error)
-      {
-         std::string msg = r::endUserErrorMessage(error);
-         reportChunkExecutionError(docId, chunkId, nbCtxId, msg, targetPath);
-         LOG_ERROR(error);
-         return Success();
-      }
-   }
-   
-   // forward success / failure to chunk
-   notebook::enqueueChunkOutput(
-            docId,
-            chunkId,
-            notebook::notebookCtxId(),
-            kChunkOutputText,
-            targetPath);
-   
-   return Success();
-}
-
-Error executeAlternateEngineChunk(const json::JsonRpcRequest& request,
-                                  json::JsonRpcResponse* pResponse)
-{
-   std::string docId, chunkId, engine, code;
-   int commitMode;
-   json::Object jsonChunkOptions;
-   Error error = json::readParams(request.params,
-                                  &docId,
-                                  &chunkId,
-                                  &commitMode,
-                                  &engine,
-                                  &code,
-                                  &jsonChunkOptions);
-   if (error)
-   {
-      LOG_ERROR(error);
-      return error;
-   }
-   
-   // read json chunk options
-   std::map<std::string, std::string> options;
-   for (json::Object::const_iterator it = jsonChunkOptions.begin();
-        it != jsonChunkOptions.end();
-        ++it)
-   {
-      options[it->first] = it->second.get_str();
-   }
-   
-   // choose appropriate notebook context to write to -- if this is a saved
-   // Rmd, we'll write directly to the saved context
-   std::string nbCtxId = 
-       static_cast<notebook::CommitMode>(commitMode) == notebook::ModeCommitted ?
-         kSavedCtx : notebook::notebookCtxId();
-
-   // handle Rcpp specially
-   if (engine == "Rcpp")
-      return executeRcppEngineChunk(docId, chunkId, nbCtxId, code, options, pResponse);
-   else if (engine == "stan")
-      return executeStanEngineChunk(docId, chunkId, nbCtxId, code, options, pResponse);
-   
-   notebook::runChunk(docId, chunkId, nbCtxId, engine, code);
-   return Success();
-}
-
-Error interruptChunk(const json::JsonRpcRequest& request,
-                     json::JsonRpcResponse* pResponse)
-{
-   std::string docId, chunkId;
-   Error error = json::readParams(request.params,
-                                  &docId,
-                                  &chunkId);
-   if (error)
-   {
-      LOG_ERROR(error);
-      return error;
-   }
-   
-   notebook::interruptChunk(docId, chunkId);
-   return Success();
-}
-
 SEXP rs_paramsFileForRmd(SEXP fileSEXP)
 {
    static std::map<std::string,std::string> s_paramsFiles;
@@ -1786,8 +1421,6 @@ Error initialize()
       (bind(registerRpcMethod, "get_rmd_template", getRmdTemplate))
       (bind(registerRpcMethod, "prepare_for_rmd_chunk_execution", prepareForRmdChunkExecution))
       (bind(registerRpcMethod, "maybe_copy_website_asset", maybeCopyWebsiteAsset))
-      (bind(registerRpcMethod, "execute_alternate_engine_chunk", executeAlternateEngineChunk))
-      (bind(registerRpcMethod, "interrupt_chunk", interruptChunk))
       (bind(registerUriHandler, kRmdOutputLocation, handleRmdOutputRequest))
       (bind(module_context::sourceModuleRFile, "SessionRMarkdown.R"));
