+++ src/cpp/session/modules/rmarkdown/SessionRmdNotebook.cpp
@@ -23,6 +23,7 @@
 #include "NotebookExec.hpp"
 #include "NotebookErrors.hpp"
 #include "NotebookQueue.hpp"
+#include "NotebookAlternateEngines.hpp"
 
 #include <iostream>
 
@@ -128,106 +129,11 @@ void emitOutputFinished(const std::string& docId, const std::string& chunkId,
    module_context::enqueClientEvent(event);
 }
 
-void onActiveConsoleChanged(const std::string& consoleId, 
-                            const std::string& text)
-{
-   s_activeConsole = consoleId;
-   if (!s_execContext)
-      return;
-
-   if (consoleId == s_execContext->chunkId())
-   {
-      if (s_execContext->connected()) 
-         return;
-      s_execContext->connect();
-      s_execContext->onConsoleInput(text);
-   }
-   else if (s_execContext->connected())
-   {
-      s_execContext->disconnect();
-      s_execContext.reset();
-   }
-}
-
 void onChunkExecCompleted(const std::string& docId, 
                           const std::string& chunkId,
                           const std::string& nbCtxId)
 {
-   // if this event belonged to the current execution context, destroy it
-   if (s_execContext &&
-       s_execContext->docId() == docId &&
-       s_execContext->chunkId() == chunkId)
-   {
-      emitOutputFinished(docId, chunkId, s_execContext->execScope());
-      s_execContext.reset();
-   }
-   else
-   {
-      emitOutputFinished(docId, chunkId, ExecScopeChunk);
-   }
-}
-
-// called by the client to set the active chunk console
-Error setChunkConsole(const json::JsonRpcRequest& request,
-                      json::JsonRpcResponse* pResponse)
-{
-   std::string docId, chunkId, options;
-   int pixelWidth = 0, charWidth = 0, execMode = 0, execScope = 0,
-       commitMode = 0;
-   Error error = json::readParams(request.params, &docId, &chunkId, &commitMode,
-         &execMode, &execScope, &options, &pixelWidth, &charWidth);
-   if (error)
-      return error;
-
-   // evaluate this chunk's options
-   r::sexp::Protect protect;
-   SEXP sexpOptions = R_NilValue;
-   error = r::exec::RFunction(".rs.evaluateChunkOptions", options)
-                             .call(&sexpOptions, &protect);
-   if (error)
-      return error;
-
-   // convert to JSON for client
-   json::Value jsonOptions;
-   error = r::json::jsonValueFromList(sexpOptions, &jsonOptions);
-   if (error)
-      return error;
-   pResponse->setResult(jsonOptions);
-
-   // if this chunk is going to be evaluated in batch mode, and the options
-   // indicate that it shouldn't be evaluated, don't
-   // evaluate it
-   if (execMode == ExecModeBatch &&
-       jsonOptions.type() == json::ObjectType)
-   {
-      bool eval = true;
-      error = json::readObject(jsonOptions.get_obj(), "eval", &eval);
-      if (!error && !eval) 
-      {
-         return Success();
-      }
-   }
-
-   // clean up the old execution context if we still have one
-   if (s_execContext)
-      s_execContext->disconnect();
-
-   // choose appropriate notebook context to write to -- if this is a saved
-   // Rmd, we'll write directly to the saved context
-   std::string nbCtxId = 
-       static_cast<CommitMode>(commitMode) == ModeCommitted ?
-         kSavedCtx :
-         notebookCtxId();
-
-   const json::Object& optionsJson = jsonOptions.type() == json::ObjectType ? 
-      jsonOptions.get_obj() : json::Object();
-   s_execContext.reset(new ChunkExecContext(docId, chunkId, nbCtxId, 
-            static_cast<ExecScope>(execScope), optionsJson, pixelWidth, 
-            charWidth));
-   if (s_activeConsole == chunkId)
-      s_execContext->connect();
-
-   return Success();
+   emitOutputFinished(docId, chunkId, ExecScopeChunk);
 }
 
 } // anonymous namespace
@@ -254,15 +160,11 @@ Error initialize()
    using boost::bind;
    using namespace module_context;
 
-   module_context::events().onActiveConsoleChanged.connect(
-         onActiveConsoleChanged);
-
    events().onChunkExecCompleted.connect(onChunkExecCompleted);
 
    ExecBlock initBlock;
    initBlock.addFunctions()
       (bind(registerRpcMethod, "refresh_chunk_output", refreshChunkOutput))
-      (bind(registerRpcMethod, "set_chunk_console", setChunkConsole))
       (bind(module_context::sourceModuleRFile, "SessionRmdNotebook.R"))
       (bind(initOutput))
       (bind(initCache))
@@ -270,7 +172,8 @@ Error initialize()
       (bind(initErrors))
       (bind(initPlots))
       (bind(initPlotReplay))
-      (bind(initQueue));
+      (bind(initQueue))
+      (bind(initAlternateEngines));
 
    return initBlock.execute();
 }
